Sửa lỗi sai về ràng buộc dữ liệu

+++ src/com/view/Student.java
@@ -6,6 +6,10 @@ import com.hibernate.*;
 import javax.swing.*;
 import javax.swing.table.DefaultTableModel;
 import java.awt.event.*;
+import java.time.DateTimeException;
+import java.time.LocalDateTime;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -30,20 +34,20 @@ public class Student extends JFrame {
     public String mssv;
 
     public DefaultTableModel tableModel = new DefaultTableModel();
-    String [] columnCourse = new String [] {"STT", "Mã môn học","Tên môn học", "Số tín chỉ","Giáo viên","Phòng học","Thứ","Ca học","Slot","Slot đã đăng kí","Đăng kí"};
-    String [] columnCourseMy = new String [] {"STT", "Mã môn học","Tên môn học", "Số tín chỉ","Giáo viên","Phòng học","Thứ","Ca học","Slot","Hủy đăng kí"};
+    String[] columnCourse = new String[]{"STT", "Mã môn học", "Tên môn học", "Số tín chỉ", "Giáo viên", "Phòng học", "Thứ", "Ca học", "Slot", "Slot đã đăng kí", "Đăng kí"};
+    String[] columnCourseMy = new String[]{"STT", "Mã môn học", "Tên môn học", "Số tín chỉ", "Giáo viên", "Phòng học", "Thứ", "Ca học", "Slot", "Hủy đăng kí"};
 
-    public Student(){
+    public Student() {
 
         txtMSSV.setEnabled(false);
         List<CourseEntity> Course = new ArrayList<>();
         Course = CourseDAO.getAllCourse();
         showListCourse(Course);
 
-        
+
         add(panel1);
         setTitle("LOG IN");
-        setSize(1500,1000);
+        setSize(1500, 1000);
 
         btnRegistration.addActionListener(new ActionListener() {
             @Override
@@ -60,7 +64,7 @@ public class Student extends JFrame {
                 setVisible(false);
             }
         });
-        
+
         btnChangePassStudent.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
@@ -138,7 +142,12 @@ public class Student extends JFrame {
                 List<JoinCourseEntity> joinList = JoinCourseDAO.getAllJoinCourse(txtMSSV.getText());
                 List<StudentEntity> studentList = StudentDAO.getInfoStudentByMSSV(txtMSSV.getText());
                 List<CourseEntity> course = CourseDAO.getAllCourse();
-                int countRegister=0;
+                List<SemesterEntity> semester = SemesterDAO.getAllSemester();
+                SemesterEntity semesterEntity = new SemesterEntity();
+                for (SemesterEntity s : semester)
+                    if (s.getType() == 1)
+                        semesterEntity = SemesterDAO.getInfoSemesterByID(s.getId());
+                int countRegister = 0;
 
                 //Đếm số cột sv muôn đăng kí
                 for (int i = 0; i < tbCourse.getRowCount(); i++) {
@@ -153,7 +162,7 @@ public class Student extends JFrame {
 
                 for (int i = 0; i < tbCourse.getRowCount(); i++) {
 
-                    if(joinList.size() + countRegister <=8) {
+                    if (joinList.size() + countRegister <= 8) {
                         JoinCourseEntity join = new JoinCourseEntity();
 //                        System.out.println(temp);
                         boolean check = Boolean.valueOf(tbCourse.getValueAt(i, 10).toString());
@@ -161,25 +170,28 @@ public class Student extends JFrame {
                         if (check == true) {
                             join.setId_course(course.get(i).getId());
                             join.setId_student(studentList.get(0).getId());
-
-                            for (JoinCourseEntity c : joinList)
-                                if (checkTimeTwoCourse(course.get(i).getId(), c.getId_course()) == false) {
-                                    checkTimeCourse = false;
-                                    CourseEntity temp = CourseDAO.getInfoCourseByID(c.getId_course());
+                            join.setDayRegistration(LocalDateTime.now());
+                            if (checkTimeRegistInSession(semesterEntity, LocalDateTime.now()) == true) {
+                                for (JoinCourseEntity c : joinList)
+                                    if (checkTimeTwoCourse(course.get(i).getId(), c.getId_course()) == false) {
+                                        checkTimeCourse = false;
+                                        CourseEntity temp = CourseDAO.getInfoCourseByID(c.getId_course());
+                                        SubjectEntity sub = SubjectDAO.getInfoSubjectByID(temp.getIdSubject());
+                                        strBuild.append("Khóa học " + sub.getSubjectName() + ": Không thể đăng kí 2 giờ trùng nhau" + "\n");
+                                        break;
+                                    }
+                                if (checkTimeCourse == true) {
+                                    boolean result = JoinCourseDAO.addJoinCourse(join);
+                                    CourseEntity temp = CourseDAO.getInfoCourseByID(join.getId_course());
                                     SubjectEntity sub = SubjectDAO.getInfoSubjectByID(temp.getIdSubject());
-                                    strBuild.append("Khóa học "+ sub.getSubjectName() + ": Không thể đăng kí 2 giờ trùng nhau"+ "\n");
-                                    break;
+                                    if (result == true)
+                                        strBuild.append("Khóa học: " + sub.getSubjectName() + " đăng kí thành công " + "\n");
                                 }
-                            if(checkTimeCourse == true){
-                                boolean result = JoinCourseDAO.addJoinCourse(join);
-                                CourseEntity temp = CourseDAO.getInfoCourseByID(join.getId_course());
-                                SubjectEntity sub = SubjectDAO.getInfoSubjectByID(temp.getIdSubject());
-                                if(result==true)
-                                    strBuild.append("Khóa học: "+sub.getSubjectName() +" đăng kí thành công "+ "\n");
+                            } else {
+                                strBuild.append("Hết hạn đăng kí " + "\n");
                             }
                         }
-                    }
-                    else{
+                    } else {
                         strBuild.append("Chỉ được đăng kí tối đa 8 môn ");
                         break;
                     }
@@ -197,28 +209,34 @@ public class Student extends JFrame {
                 StringBuilder strBuild = new StringBuilder();
                 List<JoinCourseEntity> joinList = JoinCourseDAO.getAllJoinCourse(txtMSSV.getText());
                 List<StudentEntity> studentList = StudentDAO.getInfoStudentByMSSV(txtMSSV.getText());
+                List<SemesterEntity> semester = SemesterDAO.getAllSemester();
+                SemesterEntity semesterEntity = new SemesterEntity();
+                for (SemesterEntity s : semester)
+                    if (s.getType() == 1)
+                        semesterEntity = SemesterDAO.getInfoSemesterByID(s.getId());
 //                List<CourseEntity> course = CourseDAO.getAllCourse();
                 for (int i = 0; i < tbRegistation.getRowCount(); i++) {
-                        boolean check = Boolean.valueOf(tbRegistation.getValueAt(i, 9).toString());
-                        if (check == true) {
-                            JoinCourseEntity join = new JoinCourseEntity();
-                            join.setId_course(joinList.get(i).getId_course());
-                            System.out.println(joinList.get(i).getId_course());
-                            join.setId_student(studentList.get(0).getId());
-                            System.out.println(studentList.get(0).getId());
-
-                            List<JoinCourseEntity> deleteJoin = JoinCourseDAO.getInfoJoinCourseByStudentandCourseID(join.getId_student(),join.getId_course());
-                            if(deleteJoin.size()>0) {
+                    boolean check = Boolean.valueOf(tbRegistation.getValueAt(i, 9).toString());
+                    if (check == true) {
+                        JoinCourseEntity join = new JoinCourseEntity();
+                        join.setId_course(joinList.get(i).getId_course());
+                        join.setId_student(studentList.get(0).getId());
+                        if (checkTimeRegistInSession(semesterEntity, LocalDateTime.now()) == true) {
+                            List<JoinCourseEntity> deleteJoin = JoinCourseDAO.getInfoJoinCourseByStudentandCourseID(join.getId_student(), join.getId_course());
+                            if (deleteJoin.size() > 0) {
                                 CourseEntity temp = CourseDAO.getInfoCourseByID(join.getId_course());
                                 SubjectEntity sub = SubjectDAO.getInfoSubjectByID(temp.getIdSubject());
                                 boolean result = JoinCourseDAO.deleteJoinCourse(deleteJoin.get(0).getId());
                                 if (result == true) {
                                     strBuild.append(sub.getSubjectName() + " : Xóa đăng kí thành công" + "\n");
                                 } else {
-                                    strBuild.append(sub.getSubjectName() +"Xóa đăng kí thất bại" + "\n");
+                                    strBuild.append(sub.getSubjectName() + "Xóa đăng kí thất bại" + "\n");
                                 }
                             }
+                        }else{
+                            strBuild.append("Hết hạn đăng kí " + "\n");
                         }
+                    }
                 }
                 JOptionPane.showMessageDialog(panel1, strBuild.toString(), "Thông báo", JOptionPane.DEFAULT_OPTION);
                 showListJoinCourse(JoinCourseDAO.getAllJoinCourse(txtMSSV.getText()));
@@ -227,21 +245,21 @@ public class Student extends JFrame {
         });
     }
 
-    public String showPass(String pass){
+    public String showPass(String pass) {
         String result = "";
-        for(int i=0;i<pass.length();i++)
+        for (int i = 0; i < pass.length(); i++)
             result += "*";
         return result;
     }
 
-    public void showListCourse(List<CourseEntity> list){
-        int size= list.size();
-        Object [][]course=new Object[size][11];
-        int j=0;
-        for (int i=0;i<size; i++){
+    public void showListCourse(List<CourseEntity> list) {
+        int size = list.size();
+        Object[][] course = new Object[size][11];
+        int j = 0;
+        for (int i = 0; i < size; i++) {
             SubjectEntity subjectEntity = SubjectDAO.getInfoSubjectByID(list.get(i).getIdSubject());
             SemesterEntity semesterEntity = SemesterDAO.getInfoSemesterByID(list.get(i).getIdSemester());
-            if(semesterEntity.getType()==1) {
+            if (semesterEntity.getType() == 1) {
                 course[i][0] = j + 1;
                 course[i][1] = subjectEntity.getSubjectId();
                 course[i][2] = subjectEntity.getSubjectName();
@@ -258,51 +276,73 @@ public class Student extends JFrame {
         }
 
 
-        tbCourse.setModel(new DefaultTableModel(course,columnCourse) {Class[] types = new Class[]{
-                java.lang.Integer.class, java.lang.String.class,  java.lang.String.class,
-                java.lang.String.class, java.lang.String.class, java.lang.String.class,
-                java.lang.String.class, java.lang.String.class, java.lang.Integer.class, java.lang.Integer.class, java.lang.Boolean.class};
-            public  Class getColumnClass(int columnIndex){return  types[columnIndex];}
-        }
+        tbCourse.setModel(new DefaultTableModel(course, columnCourse) {
+                              Class[] types = new Class[]{
+                                      java.lang.Integer.class, java.lang.String.class, java.lang.String.class,
+                                      java.lang.String.class, java.lang.String.class, java.lang.String.class,
+                                      java.lang.String.class, java.lang.String.class, java.lang.Integer.class, java.lang.Integer.class, java.lang.Boolean.class};
+
+                              public Class getColumnClass(int columnIndex) {
+                                  return types[columnIndex];
+                              }
+                          }
         );
     }
 
-    public boolean checkTimeTwoCourse(int idCourse1, int idCourse2)
-    {
+    public boolean checkTimeTwoCourse(int idCourse1, int idCourse2) {
         CourseEntity course1 = CourseDAO.getInfoCourseByID(idCourse1);
         CourseEntity course2 = CourseDAO.getInfoCourseByID(idCourse2);
-        if(course1 == null || course2 == null)
+        if (course1 == null || course2 == null)
             return false;
-        if(course1.getTimeOfDay().compareTo(course2.getTimeOfDay())==0 && course1.getDayOfWeek().compareTo(course2.getDayOfWeek())==0)
+        if (course1.getTimeOfDay().compareTo(course2.getTimeOfDay()) == 0 && course1.getDayOfWeek().compareTo(course2.getDayOfWeek()) == 0)
             return false;
         else
             return true;
     }
 
-    public void showListJoinCourse(List<JoinCourseEntity> list){
-        int size= list.size();
-        Object [][]course=new Object[size][10];
-        for (int i=0;i<size; i++){
+    public void showListJoinCourse(List<JoinCourseEntity> list) {
+        int size = list.size();
+        Object[][] course = new Object[size][10];
+        for (int i = 0; i < size; i++) {
             CourseEntity courseEntity = CourseDAO.getInfoCourseByID(list.get(i).getId_course());
             SubjectEntity subjectEntity = SubjectDAO.getInfoSubjectByID(courseEntity.getIdSubject());
-            course[i][0]=i+1;
-            course[i][1]=subjectEntity.getSubjectId();
-            course[i][2]=subjectEntity.getSubjectName();
-            course[i][3]=subjectEntity.getCredits();
-            course[i][4]=courseEntity.getTeacherName();
-            course[i][5]=courseEntity.getRoomName();
-            course[i][6]=courseEntity.getDayOfWeek();
-            course[i][7]=courseEntity.getTimeOfDay();
-            course[i][8]=courseEntity.getSlotMax();
-            course[i][9]=false;
+            course[i][0] = i + 1;
+            course[i][1] = subjectEntity.getSubjectId();
+            course[i][2] = subjectEntity.getSubjectName();
+            course[i][3] = subjectEntity.getCredits();
+            course[i][4] = courseEntity.getTeacherName();
+            course[i][5] = courseEntity.getRoomName();
+            course[i][6] = courseEntity.getDayOfWeek();
+            course[i][7] = courseEntity.getTimeOfDay();
+            course[i][8] = courseEntity.getSlotMax();
+            course[i][9] = false;
 
         }
 
-        tbRegistation.setModel(new DefaultTableModel(course,columnCourseMy) {Class[] types = new Class[]{
-                java.lang.Integer.class, java.lang.String.class,  java.lang.String.class,
-                java.lang.String.class, java.lang.String.class, java.lang.String.class,
-                java.lang.String.class, java.lang.String.class, java.lang.Integer.class, java.lang.Boolean.class};
-            public  Class getColumnClass(int columnIndex){return  types[columnIndex];}
+        tbRegistation.setModel(new DefaultTableModel(course, columnCourseMy) {
+            Class[] types = new Class[]{
+                    java.lang.Integer.class, java.lang.String.class, java.lang.String.class,
+                    java.lang.String.class, java.lang.String.class, java.lang.String.class,
+                    java.lang.String.class, java.lang.String.class, java.lang.Integer.class, java.lang.Boolean.class};
+
+            public Class getColumnClass(int columnIndex) {
+                return types[columnIndex];
+            }
         });
+
+    }
+
+    public boolean checkTimeRegistInSession(SemesterEntity s, LocalDateTime t) {
+        List<SessionEntity> sessionEntities = SessionDAO.getInfoSessionByIdSemester(s.getId());
+        for (SessionEntity session : sessionEntities) {
+            System.out.println(session.getDateBegin().toLocalDate().compareTo(t.toLocalDate()));
+            System.out.println(session.getDateEnd().toLocalDate().compareTo(t.toLocalDate()));
+            if (session.getDateBegin().toLocalDate().compareTo(t.toLocalDate()) <0
+                    && session.getDateEnd().toLocalDate().compareTo(t.toLocalDate()) >0) {
+                return true;
+            }
+
+        }
+        return false;
     }
 }
