+++ src/com/view/manager.java
@@ -99,6 +99,8 @@ public class manager extends JFrame {
     private JTextField txtSearchStudent;
     private JButton btnShowCourseOfStudent;
     private JButton btnShowStudentOfCourse;
+    private JButton btnSearchStudentOfCourse;
+    private JLabel lbcourse;
     public String passProfile;
 
 
@@ -109,7 +111,7 @@ public class manager extends JFrame {
     String [] columnSession = new String [] {"STT", "Tên học kì","Năm học","Ngày bắt đầu","Ngày kết thúc"};
     String [] columnCourse = new String [] {"STT", "Mã môn học","Tên môn học", "Số tín chỉ","Giáo viên","Phòng học","Thứ","Ca học","Slot","Năm học"};
     String [] columnStudent = new String [] {"STT","Lớp", "MSSV", "Mật khẩu", "Họ tên", "Email","Giới tính"};
-    String [] columnStudentOfCourse = new String [] {"STT","Lớp", "MSSV", "Họ tên", "Email","Giới tính"};
+    String [] columnStudentOfCourse = new String [] {"STT","Lớp", "MSSV", "Họ tên", "Mã môn học","Tên môn học","Tên giáo viên","Thời gian học","Thứ","Thời gian đăng kí"};
 
 
     public DefaultTableModel tableModel = new DefaultTableModel();
@@ -503,7 +505,7 @@ public class manager extends JFrame {
                 for (int i = 0; i < finalSemester.size(); i++) {
                     if (finalSemester.get(i).getSemesterName().compareToIgnoreCase((String) cmbNameSemester.getItemAt(cmbNameSemester.getSelectedIndex())) == 0
                             && finalSemester.get(i).getYear().compareTo((String) cmbYearSemester.getItemAt(cmbYearSemester.getSelectedIndex())) == 0) {
-                        strBuild.append("Lớp học đã tồn tại tài khoản.");
+                        strBuild.append("Học kì đã tồn tại tài khoản.");
                         break;
                     }
                 }
@@ -531,23 +533,44 @@ public class manager extends JFrame {
             @Override
             public void actionPerformed(ActionEvent e) {
                 List<SemesterEntity> finalSemester = SemesterDAO.getAllSemester();
-                int check = 0;
-
+                int semester = 0;
+                List<CourseEntity> courseEntities = new ArrayList<>();
+                for (SemesterEntity s : finalSemester) {
+                    if (s.getType() == 1) {
+                        courseEntities = CourseDAO.getInfoCourseBySemester(s.getId());
+                        semester=s.getId();
+                    }
+                }
                 StringBuilder strBuild = new StringBuilder();
-                if (SemesterDAO.getInfoSemesterByNameYear((String) cmbNameSemester.getItemAt(cmbNameSemester.getSelectedIndex()), (String) cmbYearSemester.getItemAt(cmbYearSemester.getSelectedIndex())).size() == 0) {
-                    strBuild.append("Học kì không tồn tại");
-                } else {
-                    List<SemesterEntity> list = SemesterDAO.getInfoSemesterByNameYear((String) cmbNameSemester.getItemAt(cmbNameSemester.getSelectedIndex()), (String) cmbYearSemester.getItemAt(cmbYearSemester.getSelectedIndex()));
-                    boolean result = SemesterDAO.deleteSemester((String) cmbNameSemester.getItemAt(cmbNameSemester.getSelectedIndex()), (String) cmbYearSemester.getItemAt(cmbYearSemester.getSelectedIndex()));
-                    if (result == true) {
-                        strBuild.append("Xóa thành công");
+                if(SemesterDAO.getInfoSemesterByNameYear((String) cmbNameSemester.getItemAt(cmbNameSemester.getSelectedIndex()),
+                        (String) cmbYearSemester.getItemAt(cmbYearSemester.getSelectedIndex())).size()>0) {
+                    if (CourseDAO.getInfoCourseBySemester(SemesterDAO.getInfoSemesterByNameYear((String) cmbNameSemester.getItemAt(cmbNameSemester.getSelectedIndex()),
+                            (String) cmbYearSemester.getItemAt(cmbYearSemester.getSelectedIndex())).get(0).getId()).size() > 0)
+                        strBuild.append("Tồn tại khóa học thuộc học kì nên không thể xóa");
+                    else {
+                        if (SemesterDAO.getInfoSemesterByNameYear((String) cmbNameSemester.getItemAt(cmbNameSemester.getSelectedIndex()), (String) cmbYearSemester.getItemAt(cmbYearSemester.getSelectedIndex())).size() == 0) {
+                            strBuild.append("Học kì không tồn tại");
+                        } else if (SemesterDAO.getInfoSemesterByNameYear((String) cmbNameSemester.getItemAt(cmbNameSemester.getSelectedIndex()), (String) cmbYearSemester.getItemAt(cmbYearSemester.getSelectedIndex())).get(0).getType() == 1) {
+                            strBuild.append("Đang là học kì hiện tại không thể xóa");
+                        } else {
+                            boolean result = false;
+                            List<SemesterEntity> list = SemesterDAO.getInfoSemesterByNameYear((String) cmbNameSemester.getItemAt(cmbNameSemester.getSelectedIndex()), (String) cmbYearSemester.getItemAt(cmbYearSemester.getSelectedIndex()));
+                            if(list.size()>0) {
+                                SessionDAO.DeleteSessionByIdSemester(list.get(0).getId());
+                                result = SemesterDAO.deleteSemester((String) cmbNameSemester.getItemAt(cmbNameSemester.getSelectedIndex()), (String) cmbYearSemester.getItemAt(cmbYearSemester.getSelectedIndex()));
+                            }
+                            if (result == true) {
+                                strBuild.append("Xóa thành công");
 
-                    } else {
-                        strBuild.append("Xóa thất bại");
+                            } else {
+                                strBuild.append("Xóa thất bại");
+                            }
+                        }
                     }
                 }
                 JOptionPane.showMessageDialog(panel1, strBuild.toString(), "Thông báo", JOptionPane.DEFAULT_OPTION);
                 showListSemester(SemesterDAO.getAllSemester());
+                showListSession(SessionDAO.getAllSession());
             }
         });
         btnSetSemester.addActionListener(new ActionListener() {
@@ -912,12 +935,15 @@ public class manager extends JFrame {
                         strBuild.append("Chọn dòng để xóa");
                     }
                     else {
-                        boolean result = CourseDAO.DeleteCourse2(course.get(check).getId());
-                        if (result == true) {
-                            strBuild.append("Xóa thành công");
-                        } else {
-                            strBuild.append("Xóa thất bại");
-                        }
+                        if (JoinCourseDAO.getAllStudentJoinCourseId(course.get(check).getId()).size() <= 0) {
+                            boolean result = CourseDAO.DeleteCourse2(course.get(check).getId());
+                            if (result == true) {
+                                strBuild.append("Xóa thành công");
+                            } else {
+                                strBuild.append("Xóa thất bại");
+                            }
+                        } else
+                            strBuild.append("Có sinh viên đăng kí nên không thể xóa");
                     }
                 JOptionPane.showMessageDialog(panel1, strBuild.toString(), "Thông báo", JOptionPane.DEFAULT_OPTION);
                 showListCourse(CourseDAO.getAllCourse());
@@ -1005,41 +1031,48 @@ public class manager extends JFrame {
             @Override
             public void actionPerformed(ActionEvent e) {
                 StringBuilder strBuild = new StringBuilder();
-                if (txtSearchStudent.getText().equals(""))
-                    strBuild.append("Ô tìm kiếm không thể để trống");
-                if (strBuild.length() > 0) {
-                    JOptionPane.showMessageDialog(panel1, strBuild.toString(), "Error", JOptionPane.ERROR_MESSAGE);
+                List<StudentEntity> student = StudentDAO.getAllStudent();
+                int check = tbStudent.getSelectedRow();
+                if (check < 0) {
+                    strBuild.append("Chọn sinh viên để xem danh sách khóa học ");
                 } else {
-                    List<StudentEntity> studentEntities = StudentDAO.getInfoStudentByMSSV(txtSearchStudent.getText());
-                    if (studentEntities.size() > 0) {
-                        List<JoinCourseEntity> joinCourseEntities = JoinCourseDAO.getAllJoinCourse(txtSearchStudent.getText());
-                        showListJoinCourse(joinCourseEntities);
-                    } else {
-                        strBuild.append("Sinh viên không tồn tại");
-                        JOptionPane.showMessageDialog(panel1, strBuild.toString(), "Thông báo", JOptionPane.DEFAULT_OPTION);
-                    }
+                    List<JoinCourseEntity> joinCourseEntities = JoinCourseDAO.getAllJoinCourse(student.get(check).getMssv());
+                    showListJoinCourse(joinCourseEntities);
                 }
+                JOptionPane.showMessageDialog(panel1, strBuild.toString(), "Thông báo", JOptionPane.DEFAULT_OPTION);
             }
         });
         btnShowStudentOfCourse.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
                 StringBuilder strBuild = new StringBuilder();
-                if(txtSearchCourse.getText().equals(""))
-                    strBuild.append("Ô tìm kiếm không thể để trống");
-                if (strBuild.length() > 0) {
-                    JOptionPane.showMessageDialog(panel1, strBuild.toString(), "Error", JOptionPane.ERROR_MESSAGE);
-                } else {
-                    List<CourseEntity> courseEntities = CourseDAO.getInfoCourseById(txtSearchCourse.getText());
-                    System.out.println(courseEntities.size());
-                    if (courseEntities.size()>0) {
-                        List<JoinCourseEntity> joinCourseEntities = JoinCourseDAO.getAllStudentJoinCourse(txtSearchStudent.getText());
-                        showListStudentOfCourse(joinCourseEntities);
-                    } else {
-                        strBuild.append("Khóa học không tồn tại");
-                        JOptionPane.showMessageDialog(panel1, strBuild.toString(), "Thông báo", JOptionPane.DEFAULT_OPTION);
+                List<CourseEntity> course = new ArrayList<>();
+                List<SemesterEntity> semester = SemesterDAO.getAllSemester();
+                for (SemesterEntity s : semester)
+                    if (s.getType() == 1) {
+                        course = CourseDAO.getInfoCourseBySemester(s.getId());
                     }
+
+                int check = tbCourse.getSelectedRow();
+                if (check < 0) {
+                    strBuild.append("Chọn khóa học để xem danh sách sinh viên");
+                } else {
+                    List<JoinCourseEntity> joinCourseEntities = JoinCourseDAO.getAllStudentJoinCourseId(course.get(check).getId());
+                    showListStudentOfCourse(joinCourseEntities);
+                    lbcourse.setText(SubjectDAO.getInfoSubjectByID(CourseDAO.getInfoCourseByID(course.get(check).getIdSubject()).getIdSubject()).getSubjectName());
                 }
+                JOptionPane.showMessageDialog(panel1, strBuild.toString(), "Thông báo", JOptionPane.DEFAULT_OPTION);
+            }
+        });
+        btnSearchStudentOfCourse.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                List<JoinCourseEntity> joinCourseEntities = new ArrayList<>();
+                if (txtSearchCourse.getText().compareTo("") != 0) {
+                    List<StudentEntity> studentEntities = StudentDAO.getInfoStudentByMSSV(txtSearchCourse.getText());
+                    joinCourseEntities = JoinCourseDAO.getAllStudentJoinCourseIdAndStudentID(CourseDAO.getInfoCourseByName(lbcourse.getText()).get(0).getId(),studentEntities.get(0).getId());
+                }
+                showListStudentOfCourse(joinCourseEntities);
             }
         });
     }
@@ -1169,34 +1202,45 @@ public class manager extends JFrame {
     public void showListJoinCourse(List<JoinCourseEntity> list) {
         int size = list.size();
         Object[][] course = new Object[size][10];
+        int j=0;
         for (int i = 0; i < size; i++) {
             CourseEntity courseEntity = CourseDAO.getInfoCourseByID(list.get(i).getId_course());
             SubjectEntity subjectEntity = SubjectDAO.getInfoSubjectByID(courseEntity.getIdSubject());
-            course[i][0] = i + 1;
-            course[i][1] = subjectEntity.getSubjectId();
-            course[i][2] = subjectEntity.getSubjectName();
-            course[i][3] = subjectEntity.getCredits();
-            course[i][4] = courseEntity.getTeacherName();
-            course[i][5] = courseEntity.getRoomName();
-            course[i][6] = courseEntity.getDayOfWeek();
-            course[i][7] = courseEntity.getTimeOfDay();
-            course[i][8] = courseEntity.getSlotMax();
+            SemesterEntity semesterEntity = SemesterDAO.getInfoSemesterByID(courseEntity.getIdSemester());
+            if(semesterEntity.getType()==1) {
+                course[i][0] = j + 1;
+                course[i][1] = subjectEntity.getSubjectId();
+                course[i][2] = subjectEntity.getSubjectName();
+                course[i][3] = subjectEntity.getCredits();
+                course[i][4] = courseEntity.getTeacherName();
+                course[i][5] = courseEntity.getRoomName();
+                course[i][6] = courseEntity.getDayOfWeek();
+                course[i][7] = courseEntity.getTimeOfDay();
+                course[i][8] = courseEntity.getSlotMax();
+                course[i][9] = semesterEntity.getSemesterName() + " " + semesterEntity.getYear();
+                j++;
+            }
         }
         tbStudent.setModel(new DefaultTableModel(course, columnCourse));
     }
 
     public void showListStudentOfCourse(List<JoinCourseEntity> list){
         int size= list.size();
-        Object [][]students=new Object[size][7];
+        Object [][]students=new Object[size][10];
         for (int i=0;i<size; i++){
             students[i][0]= i+1;
-            CourseEntity courseEntity = CourseDAO.getInfoCourseByID(list.get(i).getId_course());
-            StudentEntity studentEntity = StudentDAO.getInfoStudentByID(courseEntity.getId());
+            CourseEntity courseEntity =CourseDAO.getInfoCourseByID(list.get(0).getId_course());
+            StudentEntity studentEntity = StudentDAO.getInfoStudentByID(list.get(i).getId_student());
+            SubjectEntity subjectEntity = SubjectDAO.getInfoSubjectByID(courseEntity.getIdSubject());
             students[i][1]=studentEntity.getClassName();
             students[i][2]=studentEntity.getMssv();
-            students[i][4]=studentEntity.getFullname();
-            students[i][5]=studentEntity.getEmail();
-            students[i][6]=studentEntity.getGender();
+            students[i][3]=studentEntity.getFullname();
+            students[i][4]=subjectEntity.getSubjectId();
+            students[i][5]=subjectEntity.getSubjectName();
+            students[i][6]=courseEntity.getTeacherName();
+            students[i][7]=courseEntity.getTimeOfDay();
+            students[i][8]=courseEntity.getDayOfWeek();
+            students[i][9]=list.get(i).getDayRegistration();
         }
         tbCourse.setModel(new DefaultTableModel(students,columnStudentOfCourse));
     }
